+++ CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - `CONTAINS` edges are generated for `METHOD` to body vertices.
 - `ListMapper` to process Scala lists to a serialized string and back. More formally processing Scala lists to and from
   OverflowDB node objects.
+- Handle inheritance edges i.e. `TYPE_DECL -INHERITS_FROM-> TYPE`
 
 ### Changed